Proof-of-concept for speeding up _compute_centroid_quantile by using an accumulation tree
A profiling run shows that _compute_centroid_quantile is by far a bottleneck,
and that intuitively makes sense because its runtime is bounded by the number
of centroids in the tree O(n) (as opposed to by O(log(n)) for the other tree
operations).

Ted Dunning's original paper suggests speeding up this computation by storing
partial sums at every node in the tree. This commit does that.

In this proof-of-concept, this means trading in the C speed for FastRBTree for
the Python speed of RBTree. Still, I get an almost 50% speedup for a test use
case

    time python -c'import random; from tdigest import TDigest; TDigest().batch_update(random.random() for _ in range(10000))'

proving yet again that better algorithms beat faster implementation languages.

A followup could be to write an accumulation tree in C or Cython, which should
give the best of both worlds.